back/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flask-restx==1.3.0
Flask-CORS==5.0.0
sendgrid==6.11.0
Flask-Mail==0.10.0
python-dotenv==1.0.1